feat: add modules on the create page
