+++ src/helpers/azion-contact-sales-window-opener.js
@@ -1,0 +1,3 @@
+export const contactSalesEdgeApplicationService = () => {
+  window.open('https://www.azion.com/en/contact-sales/', '_blank')
+}
